Starting up the repository.
I started to rough out the engine: keeping track of the HTML5 canvas and
drawing contexts, introducing some globals and working on a an Actor
class (basically a sprite class). Additionally, I've included the source
to the first game I'll be making for my game a week challenge (although
the first one may take two weeks since I'll be writing all of the basic
stuff that the engine will need for this and future games.

P.S. GIT's windows native app is pretty sweet.