Introducing some basic physics functions
Started working on the physics portion of the game/game engine. So far
I've written a VERY basic kinematics function and created a new version
of PenultActor that should be usable with the physics engine later on.
Right now the PenultPhysics class is little more than utility functions.
Additionally I've applied a refactor to the PenultActor, because I was
really unhappy with the non-descriptive names I gave some of the
attributes.