Introducing Layers
So the canvas draws as though it's a display which is nice, but it can
be confusing in the context of games and when dealing with coordinate
spaces. As such I've consolidated the canvas and drawing context in an
object I've dubbed the "PenultLayer" which will serve as a go between
for drawing functions to deal with issues of coordinate space
translation. At the moment on my trial game I am hitting serious frame
rate drops, but I need to investigate more rigorously to determine the
source of the drop.